[XLA] Do fewer hashtable lookups in HloCostAnalysis.
HloCostAnalysis is slow principally because every property lookup is a lookup
in an absl_hash_map<string, float>.  These lookups are *everywhere*.

Speed here has only recently become important, now that GPU compilation uses
the cost model during fusion.

*Most* of the time, we're looking up one of a few well-defined fields.  But
sometimes we're looking up a dynamic field.

So...gods forgive me, I modified HloCostAnalysis::Properties to *act* like a
hashtable, but have a fastpath for common field names.

In fact, now properties[kBytesAccessedKey] gets optimized to be the same as
reading properties.bytes_accessed() directly.

DO NOT SUBMIT - Should I remove the bytes_accessed() accessor?  Seems
unnecessary if I claim that reading through operator[] is fast.  But...are we
100% sure?  I could mark it as always_inline?

This is a lot uglier, but it's also a lot faster.

PiperOrigin-RevId: 492655794